PWM Duty Cycle Control using POT
PWM perios has been changed such that a 12-bit value  maps directly to the duty cycle period.

A task samples and waits for a single conversion to end on ADC1_IN1@PA1 which then gets added to a global variable before averaging. This is done to obtain a more accurate and stable reading.

Another task then read the global variable periodically to update the duty cycle on TIM4@CH2@PB7. Not the best solution but just for fun. This enables testing the Sharp IR distance sensor amd the TI motor controller. See design.txt.