Motor Speed Adjust Using POT
Introduced the use of a queue to share pot sample values between the sampling and the speed control task. Queue is passed to task via task parameters. 

Plenty of globals remain which would be slowly reduced to lower scope i.e. created/init in main() and/or passed via parameters to respective functions.